+++ oceantracker/field_group_manager/setup_reader.py
@@ -109,7 +109,7 @@ def _standard_needed_info(reader):
 
     # set default z dim info to that for 2D, add_hindcast_info changes them for  3D
     info['z_dim'] = None
-    info['num_z_levels'] = 1
+    info['num_z_interfaces'] = 1
     info['all_z_dims'] = []
     info['vert_grid_type'] = None
 
@@ -292,7 +292,7 @@ def _catalog_fields(reader, crumbs=None):
 
         field_params = dict(time_varying=file_vars[v1]['time_varying'],
                             is3D=any(x in info['all_z_dims'] for x in file_vars[v1]['dims']),  )
-        field_params['zlevels'] = info['num_z_levels'] if field_params['is3D'] else 1
+        field_params['zlevels'] = info['num_z_interfaces'] if field_params['is3D'] else 1
 
         # work out if variable is a vector field
         file_vars_info = {}
@@ -310,7 +310,7 @@ def _catalog_fields(reader, crumbs=None):
 
             s4D = [si.settings.time_buffer_size if field_params['time_varying'] else 1,
                    info['num_nodes'],
-                   info['num_z_levels'] if field_params['is3D'] else 1,
+                   info['num_z_interfaces'] if field_params['is3D'] else 1,
                    n_comp]
 
             file_vars_info[v] = dict(vector_components_per_file_var=n_comp,